Update client router

+++ main-service/src/api/client.router.ts
@@ -7,13 +7,18 @@ import {
   updateClientById,
   deleteClientById,
 } from "../service/client.service";
-import { validate } from "./middleware";
+import {
+  checkIfAdmin,
+  checkIfAdminOrProjectManager,
+  validate,
+} from "./middleware";
 
 export const clientRouter = Router();
 
 clientRouter.post(
   "/",
   validate(clientSchema),
+  checkIfAdmin(),
   async (req: Request, res: Response) => {
     const payload = req.body as ClientSchema;
     const client = await createClient(payload);
@@ -22,21 +27,30 @@ clientRouter.post(
   }
 );
 
-clientRouter.get("/", async (req: Request, res: Response) => {
-  const clients = await getAllClients();
-  if (!clients) return res.sendStatus(404);
-  return res.status(200).send(clients);
-});
+clientRouter.get(
+  "/",
+  checkIfAdminOrProjectManager(),
+  async (req: Request, res: Response) => {
+    const clients = await getAllClients();
+    if (!clients) return res.sendStatus(404);
+    return res.status(200).send(clients);
+  }
+);
 
-clientRouter.get("/:id", async (req: Request, res: Response) => {
-  const clientId = req.params.id;
-  const client = await getClientById(clientId);
-  if (!client) return res.sendStatus(404);
-  return res.status(200).send(client);
-});
+clientRouter.get(
+  "/:id",
+  checkIfAdminOrProjectManager(),
+  async (req: Request, res: Response) => {
+    const clientId = req.params.id;
+    const client = await getClientById(clientId);
+    if (!client) return res.sendStatus(404);
+    return res.status(200).send(client);
+  }
+);
 
 clientRouter.patch(
   "/:id",
+  checkIfAdmin(),
   validate(clientSchema),
   async (req: Request, res: Response) => {
     const clientId = req.params.id;
@@ -47,9 +61,13 @@ clientRouter.patch(
   }
 );
 
-clientRouter.delete("/:id", async (req: Request, res: Response) => {
-  const clientId = req.params.id;
-  const client = await deleteClientById(clientId);
-  if (!client) return res.sendStatus(404);
-  return res.sendStatus(200);
-});
+clientRouter.delete(
+  "/:id",
+  checkIfAdmin(),
+  async (req: Request, res: Response) => {
+    const clientId = req.params.id;
+    const client = await deleteClientById(clientId);
+    if (!client) return res.sendStatus(404);
+    return res.sendStatus(200);
+  }
+);
